backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy[asyncio]==2.0.48
asyncpg==0.31.0
alembic==1.18.4
langchain==1.2.12
langchain-openai==1.1.11
pydantic-settings==2.13.1
python-dotenv==1.2.1
httpx==0.28.1
PyJWT==2.12.0
cryptography==46.0.5
slowapi==0.1.9
pytest==9.0.2
pytest-asyncio==1.3.0
aiosqlite==0.22.1
anyio==4.12.1
bcrypt==4.0.1
tavily-python==0.7.23
exa_py==2.9.0
duckduckgo-search==8.1.1
passlib[bcrypt]==1.7.4
argon2-cffi==25.1.0
tenacity==9.1.4
resend==2.25.0
email-validator==2.3.0
